growthleads_assignment/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
certifi==2024.6.2
    # via requests
charset-normalizer==3.3.2
    # via requests
greenlet==3.0.3
    # via sqlalchemy
idna==3.7
    # via requests
psycopg2==2.9.9
    # via assignment (setup.py)
python-restcountries==2.0.0
    # via assignment (setup.py)
requests==2.32.3
    # via python-restcountries
sqlalchemy==2.0.31
    # via assignment (setup.py)
typing-extensions==4.12.2
    # via sqlalchemy
urllib3==2.2.2
    # via requests
